+++ recorder/stream.go
@@ -2,7 +2,6 @@ package recorder
 
 import (
 	"encoding/binary"
-	"fmt"
 	"io"
 	"sync"
 	"sync/atomic"
@@ -44,19 +43,11 @@ func (q *QuicStream) Write(p []byte) (int, error) {
 		packetType = DataPacket
 	}
 
-	fmt.Println("Sending packet type", packetType)
-	fmt.Println("Original payload length", len(p))
-
 	header := make([]byte, 3)
 	header[0] = byte(packetType)
 	binary.BigEndian.PutUint16(header[1:], uint16(len(p)))
 
 	packet := append(header, p...)
-
-	fmt.Println("Sending packet with length ", len(packet), "with header length", len(header), "and payload length", uint16(len(p)))
-
-	fmt.Println("Sending packet", string(packet))
-
 	n, err := q.quicStream.Write(packet)
 	if err != nil {
 		return 0, err
